miri: improve errors for type validity assertion failures
